彩种投注 算法结果接入

+++ module-lottery/src/main/java/com/xtree/lottery/ui/lotterybet/viewmodel/LotteryBetConfirmViewModel.java
@@ -83,7 +83,7 @@ public class LotteryBetConfirmViewModel extends BaseViewModel<LotteryRepository>
 
         datas.set(bindModels);
 
-        int money = 0;
+        double money = 0;
         boolean solo = false;
         for (LotteryBetRequest.BetOrderData betOrderData : betList) {
             money += betOrderData.getMoney();
